Mostly finished refactoring. Need to check async task handler again.
