+++ src/ghga_datasteward_kit/s3_upload/entrypoint.py
@@ -19,6 +19,7 @@
 import asyncio
 import base64
 import logging
+import math
 from pathlib import Path
 
 import typer
@@ -26,39 +27,167 @@ from ghga_service_commons.utils.crypt import encrypt
 
 from ghga_datasteward_kit import models
 from ghga_datasteward_kit.s3_upload.config import Config, LegacyConfig
+from ghga_datasteward_kit.s3_upload.exceptions import (
+    MultipartUploadCompletionError,
+    SecretExchangeError,
+    WritingOutputError,
+)
 from ghga_datasteward_kit.s3_upload.file_decryption import Decryptor
 from ghga_datasteward_kit.s3_upload.file_encryption import Encryptor
+from ghga_datasteward_kit.s3_upload.http_client import RequestConfigurator, httpx_client
+from ghga_datasteward_kit.s3_upload.multipart_upload import MultipartUpload
 from ghga_datasteward_kit.s3_upload.uploader import ChunkedUploader
 from ghga_datasteward_kit.s3_upload.utils import (
     LOG,
-    RequestConfigurator,
-    SecretExchangeError,
-    StorageCleaner,
-    WritingOutputError,
     check_adjust_part_size,
     check_output_path,
     get_bucket_id,
     handle_superficial_error,
-    httpx_client,
 )
 from ghga_datasteward_kit.utils import STEWARD_TOKEN, load_config_yaml, path_join
 
 
-async def validate_and_transfer_content(
+def main(
+    input_path: Path = typer.Option(..., help="Local path of the input file"),
+    alias: str = typer.Option(..., help="A human readable file alias"),
+    config_path: Path = typer.Option(..., help="Path to a config YAML."),
+):
+    """
+    Custom script to encrypt data using Crypt4GH and directly uploading it to S3
+    object storage.
+    """
+    config = load_config_yaml(config_path, Config)
+
+    token = STEWARD_TOKEN.read_token()
+    asyncio.run(
+        async_main(input_path=input_path, alias=alias, config=config, token=token)
+    )
+
+
+def legacy_main(
+    input_path: Path = typer.Option(..., help="Local path of the input file"),
+    alias: str = typer.Option(..., help="A human readable file alias"),
+    config_path: Path = typer.Option(..., help="Path to a config YAML."),
+):
+    """
+    Custom script to encrypt data using Crypt4GH and directly uploading it to S3
+    object storage.
+    """
+    config = load_config_yaml(config_path, LegacyConfig)
+    asyncio.run(legacy_async_main(input_path=input_path, alias=alias, config=config))
+
+
+async def async_main(input_path: Path, alias: str, config: Config, token: str):
+    """
+    Run encryption, upload and validation.
+    Prints metadata to <alias>.json in the specified output directory
+    """
+    RequestConfigurator.configure(config=config)
+    file_size = await check_adjust_input_file(
+        input_path=input_path, alias=alias, config=config
+    )
+
+    async with MultipartUpload(file_size=file_size, config=config) as upload:
+        uploader = await validate_and_transfer_content(
+            input_path=input_path, config=config, upload=upload
+        )
+
+        (
+            unencrypted_checksum,
+            encrypted_md5_checksums,
+            encrypted_sha256_checksums,
+        ) = uploader.encryptor.checksums.get()
+
+        secret_id = await exchange_secret_for_id(
+            file_id=upload.file_id,
+            secret=uploader.encryptor.file_secret,
+            token=token,
+            config=config,
+        )
+
+        metadata = models.OutputMetadata(
+            alias=alias,
+            file_id=upload.file_id,
+            bucket_id=get_bucket_id(config=config),
+            object_id=upload.file_id,
+            original_path=input_path,
+            part_size=config.part_size,
+            secret_id=secret_id,
+            unencrypted_checksum=unencrypted_checksum,
+            encrypted_md5_checksums=encrypted_md5_checksums,
+            encrypted_sha256_checksums=encrypted_sha256_checksums,
+            unencrypted_size=file_size,
+            encrypted_size=upload.encrypted_file_size,
+            storage_alias=config.selected_storage_alias,
+        )
+        output_path = config.output_dir / f"{alias}.json"
+        LOG.info("(4/4) Writing metadata to %s.", output_path)
+        try:
+            metadata.serialize(output_path)
+        except (
+            Exception,
+            KeyboardInterrupt,
+        ) as exc:
+            raise WritingOutputError(
+                bucket_id=get_bucket_id(config), object_id=upload.file_id
+            ) from exc
+
+
+async def legacy_async_main(input_path: Path, alias: str, config: LegacyConfig):
+    """
+    Run encryption, upload and validation.
+    Prints metadata to <alias>.json in the specified output directory
+    """
+    RequestConfigurator.configure(config=config)
+    file_size = await check_adjust_input_file(
+        input_path=input_path, alias=alias, config=config
+    )
+
+    async with MultipartUpload(file_size=file_size, config=config) as upload:
+        uploader = await validate_and_transfer_content(
+            input_path=input_path, config=config, upload=upload
+        )
+
+        (
+            unencrypted_checksum,
+            encrypted_md5_checksums,
+            encrypted_sha256_checksums,
+        ) = uploader.encryptor.checksums.get()
+
+        file_secret = base64.b64encode(uploader.encryptor.file_secret).decode("utf-8")
+
+        metadata = models.LegacyOutputMetadata(
+            alias=alias,
+            file_id=upload.file_id,
+            bucket_id=get_bucket_id(config=config),
+            object_id=upload.file_id,
+            original_path=input_path,
+            part_size=config.part_size,
+            file_secret=file_secret,
+            unencrypted_checksum=unencrypted_checksum,
+            encrypted_md5_checksums=encrypted_md5_checksums,
+            encrypted_sha256_checksums=encrypted_sha256_checksums,
+            unencrypted_size=file_size,
+            encrypted_size=upload.encrypted_file_size,
+            storage_alias=config.selected_storage_alias,
+        )
+        output_path = config.output_dir / f"{alias}.json"
+        LOG.info("(4/4) Writing metadata to %s.", output_path)
+        try:
+            metadata.serialize(output_path)
+        except (
+            Exception,
+            KeyboardInterrupt,
+        ) as exc:
+            raise WritingOutputError(
+                bucket_id=get_bucket_id(config), object_id=upload.file_id
+            ) from exc
+
+
+async def check_adjust_input_file(
     input_path: Path, alias: str, config: LegacyConfig
-) -> tuple[ChunkedUploader, int]:
-    """
-    Check and upload encrypted file content.
-
-    This also includes a verification of the upload by calculating and supplying part MD5
-    sums and checking the ETag of the uploaded object.
-    Additionally the encrypted parts are decrypted locally and the checksum of the initial
-    unencrypted and the decrypted file are compared.
-
-
-    Returns:
-        A tuple of the used uploader instance and the file size
-    """
+) -> int:
+    """Check if input file exists, get file size and adjust part size if necessary."""
     if not input_path.exists():
         msg = f"No such file: {input_path.resolve()}"
         handle_superficial_error(msg=msg)
@@ -72,20 +201,55 @@ async def validate_and_transfer_content(
     file_size = input_path.stat().st_size
     check_adjust_part_size(config=config, file_size=file_size)
 
+    return file_size
+
+
+async def validate_and_transfer_content(
+    input_path: Path, upload: MultipartUpload, config: LegacyConfig
+) -> ChunkedUploader:
+    """
+    Check and upload encrypted file content.
+
+    This also includes a verification of the upload by calculating and supplying part MD5
+    sums and checking the ETag of the uploaded object.
+    Additionally the encrypted parts are decrypted locally and the checksum of the initial
+    unencrypted and the decrypted file are compared.
+
+
+    Returns:
+        A tuple of the used uploader instance and the file size
+    """
     encryptor = Encryptor(config.part_size)
     decryptor = Decryptor(file_secret=encryptor.file_secret)
 
     uploader = ChunkedUploader(
         input_path=input_path,
-        alias=alias,
         config=config,
-        unencrypted_file_size=file_size,
         encryptor=encryptor,
         decryptor=decryptor,
+        upload=upload,
     )
     await uploader.encrypt_and_upload()
+    try:
+        await upload.storage.complete_multipart_upload(
+            upload_id=upload.upload_id,
+            bucket_id=upload.bucket_id,
+            object_id=upload.file_id,
+            anticipated_part_quantity=math.ceil(
+                upload.encrypted_file_size / upload.part_size
+            ),
+            anticipated_part_size=upload.part_size,
+        )
+    except BaseException as exc:
+        raise MultipartUploadCompletionError(
+            cause=str(exc),
+            bucket_id=upload.bucket_id,
+            object_id=upload.file_id,
+            upload_id=upload.upload_id,
+        ) from exc
+    await upload.check_md5_matches()
 
-    return uploader, file_size
+    return uploader
 
 
 async def exchange_secret_for_id(
@@ -123,141 +287,6 @@ async def exchange_secret_for_id(
                 bucket_id=get_bucket_id(config), object_id=file_id, message=message
             )
         return response.json()["secret_id"]
-
-
-async def async_main(input_path: Path, alias: str, config: Config, token: str):
-    """
-    Run encryption, upload and validation.
-    Prints metadata to <alias>.json in the specified output directory
-    """
-    RequestConfigurator.configure(config=config)
-
-    async with StorageCleaner(config=config):
-        uploader, file_size = await validate_and_transfer_content(
-            input_path=input_path,
-            alias=alias,
-            config=config,
-        )
-
-        (
-            unencrypted_checksum,
-            encrypted_md5_checksums,
-            encrypted_sha256_checksums,
-        ) = uploader.encryptor.checksums.get()
-
-        secret_id = await exchange_secret_for_id(
-            file_id=uploader.file_id,
-            secret=uploader.encryptor.file_secret,
-            token=token,
-            config=config,
-        )
-
-        metadata = models.OutputMetadata(
-            alias=uploader.alias,
-            file_id=uploader.file_id,
-            bucket_id=get_bucket_id(config=config),
-            object_id=uploader.file_id,
-            original_path=input_path,
-            part_size=config.part_size,
-            secret_id=secret_id,
-            unencrypted_checksum=unencrypted_checksum,
-            encrypted_md5_checksums=encrypted_md5_checksums,
-            encrypted_sha256_checksums=encrypted_sha256_checksums,
-            unencrypted_size=file_size,
-            encrypted_size=uploader.encryptor.encrypted_file_size,
-            storage_alias=config.selected_storage_alias,
-        )
-        output_path = config.output_dir / f"{uploader.alias}.json"
-        LOG.info("(4/4) Writing metadata to %s.", output_path)
-        try:
-            metadata.serialize(output_path)
-        except (
-            Exception,
-            KeyboardInterrupt,
-        ) as exc:
-            raise WritingOutputError(
-                bucket_id=get_bucket_id(config), object_id=uploader.file_id
-            ) from exc
-
-
-async def legacy_async_main(input_path: Path, alias: str, config: LegacyConfig):
-    """
-    Run encryption, upload and validation.
-    Prints metadata to <alias>.json in the specified output directory
-    """
-    RequestConfigurator.configure(config=config)
-
-    async with StorageCleaner(config=config):
-        uploader, file_size = await validate_and_transfer_content(
-            input_path=input_path,
-            alias=alias,
-            config=config,
-        )
-
-        (
-            unencrypted_checksum,
-            encrypted_md5_checksums,
-            encrypted_sha256_checksums,
-        ) = uploader.encryptor.checksums.get()
-
-        file_secret = base64.b64encode(uploader.encryptor.file_secret).decode("utf-8")
-
-        metadata = models.LegacyOutputMetadata(
-            alias=uploader.alias,
-            file_id=uploader.file_id,
-            bucket_id=get_bucket_id(config=config),
-            object_id=uploader.file_id,
-            original_path=input_path,
-            part_size=config.part_size,
-            file_secret=file_secret,
-            unencrypted_checksum=unencrypted_checksum,
-            encrypted_md5_checksums=encrypted_md5_checksums,
-            encrypted_sha256_checksums=encrypted_sha256_checksums,
-            unencrypted_size=file_size,
-            encrypted_size=uploader.encryptor.encrypted_file_size,
-            storage_alias=config.selected_storage_alias,
-        )
-        output_path = config.output_dir / f"{uploader.alias}.json"
-        LOG.info("(4/4) Writing metadata to %s.", output_path)
-        try:
-            metadata.serialize(output_path)
-        except (
-            Exception,
-            KeyboardInterrupt,
-        ) as exc:
-            raise WritingOutputError(
-                bucket_id=get_bucket_id(config), object_id=uploader.file_id
-            ) from exc
-
-
-def main(
-    input_path: Path = typer.Option(..., help="Local path of the input file"),
-    alias: str = typer.Option(..., help="A human readable file alias"),
-    config_path: Path = typer.Option(..., help="Path to a config YAML."),
-):
-    """
-    Custom script to encrypt data using Crypt4GH and directly uploading it to S3
-    object storage.
-    """
-    config = load_config_yaml(config_path, Config)
-
-    token = STEWARD_TOKEN.read_token()
-    asyncio.run(
-        async_main(input_path=input_path, alias=alias, config=config, token=token)
-    )
-
-
-def legacy_main(
-    input_path: Path = typer.Option(..., help="Local path of the input file"),
-    alias: str = typer.Option(..., help="A human readable file alias"),
-    config_path: Path = typer.Option(..., help="Path to a config YAML."),
-):
-    """
-    Custom script to encrypt data using Crypt4GH and directly uploading it to S3
-    object storage.
-    """
-    config = load_config_yaml(config_path, LegacyConfig)
-    asyncio.run(legacy_async_main(input_path=input_path, alias=alias, config=config))
 
 
 if __name__ == "__main__":
